+++ src/main/java/com/gymtracker/app/entity/workout/WorkoutRepetitionItemEntity.java
@@ -12,6 +12,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Getter
@@ -33,6 +34,6 @@ public class WorkoutRepetitionItemEntity extends WorkoutItemEntity {
     @AllArgsConstructor
     public static class ExerciseSet {
         private int reps;
-        private double weight;
+        private BigDecimal weight;
     }
 }
